RavenDB-20799 Preparing view
