+++ src/Raven.Studio/typescript/components/pages/database/tasks/ongoingTasks/OngoingTaskPage.scss
