+++ src/Raven.Studio/typescript/components/pages/database/tasks/ongoingTasks/OngoingTasksPage.spec.tsx
