+++ src/Raven.Studio/typescript/components/pages/database/tasks/ongoingTasks/OngoingTasksReducer.ts
