+++ src/Raven.Studio/typescript/components/pages/database/tasks/ongoingTasks/panels/RavenEtlPanel.tsx
@@ -1,5 +1,5 @@
-﻿import React from "react";
-import { OngoingTaskExternalReplicationInfo } from "components/models/tasks";
+﻿import React, { useCallback } from "react";
+import { useAccessManager } from "hooks/useAccessManager";
 import {
     RichPanel,
     RichPanelActions,
@@ -7,36 +7,33 @@ import {
     RichPanelDetails,
     RichPanelHeader,
     RichPanelInfo,
+    RichPanelSelect,
 } from "components/common/RichPanel";
 import {
     ConnectionStringItem,
+    EmptyScriptsWarning,
+    ICanShowTransformationScriptPreview,
     OngoingTaskActions,
     OngoingTaskName,
     OngoingTaskResponsibleNode,
     OngoingTaskStatus,
-} from "../shared";
-import { useAccessManager } from "hooks/useAccessManager";
+} from "../../shared";
 import { useAppUrls } from "hooks/useAppUrls";
-import { BaseOngoingTaskPanelProps, useTasksOperations } from "../shared";
-import genUtils from "common/generalUtils";
-import { Collapse } from "reactstrap";
+import { OngoingTaskRavenEtlInfo } from "components/models/tasks";
+import { BaseOngoingTaskPanelProps, useTasksOperations } from "../../shared";
+import { OngoingEtlTaskDistribution } from "./OngoingEtlTaskDistribution";
+import { Collapse, Input } from "reactstrap";
 
-type ExternalReplicationPanelProps = BaseOngoingTaskPanelProps<OngoingTaskExternalReplicationInfo>;
+type RavenEtlPanelProps = BaseOngoingTaskPanelProps<OngoingTaskRavenEtlInfo>;
 
-function Details(props: ExternalReplicationPanelProps & { canEdit: boolean }) {
+function Details(props: RavenEtlPanelProps & { canEdit: boolean }) {
     const { data, canEdit, db } = props;
-
-    const showDelayReplication = data.shared.delayReplicationTime > 0;
-    const delayHumane = genUtils.formatTimeSpan(1000 * (data.shared.delayReplicationTime ?? 0), true);
     const connectionStringDefined = !!data.shared.destinationDatabase;
     const { appUrl } = useAppUrls();
     const connectionStringsUrl = appUrl.forConnectionStrings(db, "Raven", data.shared.connectionStringName);
 
     return (
         <RichPanelDetails>
-            {showDelayReplication && (
-                <RichPanelDetailItem label="Replication Delay Time">{delayHumane}</RichPanelDetailItem>
-            )}
             <ConnectionStringItem
                 connectionStringDefined={!!data.shared.destinationDatabase}
                 canEdit={canEdit}
@@ -62,28 +59,39 @@ function Details(props: ExternalReplicationPanelProps & { canEdit: boolean }) {
                     {data.shared.topologyDiscoveryUrls.join(", ")}
                 </RichPanelDetailItem>
             )}
+            <EmptyScriptsWarning task={data} />
         </RichPanelDetails>
     );
 }
 
-export function ExternalReplicationPanel(props: ExternalReplicationPanelProps) {
-    const { db, data } = props;
+export function RavenEtlPanel(props: RavenEtlPanelProps & ICanShowTransformationScriptPreview) {
+    const { db, data, showItemPreview } = props;
 
     const { isAdminAccessOrAbove } = useAccessManager();
     const { forCurrentDatabase } = useAppUrls();
 
     const canEdit = isAdminAccessOrAbove(db) && !data.shared.serverWide;
-    const editUrl = forCurrentDatabase.editExternalReplication(data.shared.taskId)();
+    const editUrl = forCurrentDatabase.editRavenEtl(data.shared.taskId)();
 
     const { detailsVisible, toggleDetails, toggleStateHandler, onEdit, onDeleteHandler } = useTasksOperations(
         editUrl,
         props
     );
 
+    const showPreview = useCallback(
+        (transformationName: string) => {
+            showItemPreview(data, transformationName);
+        },
+        [data, showItemPreview]
+    );
+
     return (
         <RichPanel>
             <RichPanelHeader>
                 <RichPanelInfo>
+                    <RichPanelSelect>
+                        <Input type="checkbox" onChange={() => null} checked={false} />
+                    </RichPanelSelect>
                     <OngoingTaskName task={data} canEdit={canEdit} editUrl={editUrl} />
                 </RichPanelInfo>
                 <RichPanelActions>
@@ -100,6 +108,7 @@ export function ExternalReplicationPanel(props: ExternalReplicationPanelProps) {
             </RichPanelHeader>
             <Collapse isOpen={detailsVisible}>
                 <Details {...props} canEdit={canEdit} />
+                <OngoingEtlTaskDistribution task={data} showPreview={showPreview} />
             </Collapse>
         </RichPanel>
     );
