+++ src/Raven.Studio/typescript/components/pages/database/tasks/ongoingTasks/panels/ReplicationHubConnectedSinkPanel.tsx
@@ -8,8 +8,10 @@ import {
     RichPanelDetails,
     RichPanelHeader,
     RichPanelInfo,
+    RichPanelSelect,
 } from "components/common/RichPanel";
-import { OngoingTaskName, OngoingTaskResponsibleNode } from "../shared";
+import { OngoingTaskName, OngoingTaskResponsibleNode } from "../../shared";
+import { Input } from "reactstrap";
 
 interface ReplicationHubConnectedSinkPanelProps {
     db: database;
@@ -23,6 +25,9 @@ export function ReplicationHubConnectedSinkPanel(props: ReplicationHubConnectedS
         <RichPanel>
             <RichPanelHeader>
                 <RichPanelInfo>
+                    <RichPanelSelect>
+                        <Input type="checkbox" onChange={() => null} checked={false} />
+                    </RichPanelSelect>
                     <OngoingTaskName task={data} canEdit={false} editUrl={undefined} />
                 </RichPanelInfo>
                 <RichPanelActions>
